new file:   handin_1/dk/itu/bmds/teamkolera/src/lib/Connection.java 	new file:   handin_1/dk/itu/bmds/teamkolera/src/lib/Marshall.java 	new file:   handin_1/dk/itu/bmds/teamkolera/src/proj_main/TaskManagerTCPClient.java 	new file:   handin_1/dk/itu/bmds/teamkolera/src/proj_main/TaskManagerTCPServer.java 	new file:   handin_1/dk/itu/bmds/teamkolera/src/taskmanager/Attendant.java 	new file:   handin_1/dk/itu/bmds/teamkolera/src/taskmanager/Calendar.java 	new file:   handin_1/dk/itu/bmds/teamkolera/src/taskmanager/Task.java 	new file:   handin_1/dk/itu/bmds/teamkolera/src/taskmanager/User.java 	new file:   handin_1/resources/task-manager-xml.xml
The namespaces/packagenames are a littele weird, but they should make
sense.
The logic is still not imnplemented, though the "framework" should be
in place, save for error handling and tests.
(Testing and error handlig will probably never be implemented)
The post method seems complete. Will have to make sure though. Need
coffee.